Move TC resolver to its own package
